#143799299 Admin wants to add custom detailed map snippets for round locations on admin panel
Added linkify for future linking business, fixed regex(as much as i was arsed, since the original neither worked or captured the actual data we needed it to, the fact that the link is indeed an image. )
Refactored some of the REST calls to use the common makeCall, added a stacktrace function, added a class to differentiate between a round location and a general location. Haven't yet fixed the map bug with round location.